Move Node struct to modelardb_types
